@@ -3,16 +3,21 @@
   "name": "react-typescript-starter-pack",
   "version": "0.1.0",
   "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.5.1",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.10",
     "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react-dom": "^18.2.18",
+    "@types/react-transition-group": "^4.4.10",
+    "bulma": "^0.9.4",
+    "classnames": "^2.5.1",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-scripts": "4.0.3",
+    "react-transition-group": "^4.4.5",
     "typescript": "^4.2.4",
     "typescript-eslint-parser": "^22.0.0",
     "web-vitals": "^1.1.1"
