@@ -1,11 +1,15 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.scss';
-import { App } from './App';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root'),
-);
+import 'bulma/css/bulma.css';
+import '@fortawesome/fontawesome-free/css/all.css';
+import './styles/index.scss';
+
+import { App } from './App';
+import { TodosProvider } from './stores/TodosContext';
+
+createRoot(document.getElementById('root') as HTMLDivElement)
+  .render(
+    <TodosProvider>
+      <App />
+    </TodosProvider>,
+  );
